+++ ng-workspace/projects/OpenWaterFoundation/common/ui/core/chart/chart.service.ts
@@ -112,7 +112,8 @@ export class ChartService {
     let endDate: DateTime = timeSeries.getDate2();
     // The DateTime iterator for the the while loop.
     let iter: DateTime = startDate;
-    // The index of the x_axisLabel array to push into the chartJS_yAxisData as the x property.
+    // The index of the x_axisLabel array to push into the chartJS_yAxisData as
+    // the x property.
     var labelIndex = 0;      
     
     do {