+++ ng-workspace/projects/OpenWaterFoundation/common/util/datastore/datastore.spec.ts
@@ -1,11 +1,15 @@
 import { HttpClientTestingModule,
-          HttpTestingController } from '@angular/common/http/testing';
-import { TestBed }                from '@angular/core/testing';
+          HttpTestingController }         from '@angular/common/http/testing';
+import { TestBed }                        from '@angular/core/testing';
+import { HttpClient }                     from '@angular/common/http';
 
-import { DatastoreManager }       from './DataStoreManager';
-import { DateValueDatastore }     from './DateValueDataStore';
-import { DelimitedDatastore }     from './DelimitedDataStore';
-import { StateModDatastore }      from './StateModDataStore';
+import { OwfCommonService }               from '@OpenWaterFoundation/common/services';
+
+import { DatastoreManager }               from './DataStoreManager';
+import { DateValueDatastore }             from './DateValueDataStore';
+import { DelimitedDatastore }             from './DelimitedDataStore';
+import { StateModDatastore }              from './StateModDataStore';
+import { ColoradoHydroBaseRestDatastore } from './ColoradoHydroBaseRestDatastore';
 
 
 xdescribe('Datastores', () => {
@@ -13,7 +17,7 @@ xdescribe('Datastores', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule]
+      imports: [HttpClient]
     });
     httpTestingController = TestBed.inject(HttpTestingController);
   });
@@ -25,6 +29,15 @@ xdescribe('Datastores', () => {
   xdescribe('when using the DateValue datastore', () => {
 
     it('should correctly convert a path', () => {
+      
+    });
+  });
+
+  xdescribe('when using the ColoradoHydroBaseRest datastore', () => {
+
+    // var hydroBaseDatastore = new ColoradoHydroBaseRestDatastore();
+
+    it('should asynchronously get Telemetry Parameters on instance creation', () => {
       
     });
   });