+++ ng-workspace/projects/OpenWaterFoundation/common/CHANGELOG.md
@@ -10,7 +10,7 @@ Optional elements to be added to each package version are as follows:
 
 These elements will only be added if they are applicable for the new version.
 
-# 3.6.2 or 3.7.0 #
+# 3.7.0 #
 
 ### Refactoring ###
 
@@ -18,6 +18,9 @@ These elements will only be added if they are applicable for the new version.
 
 ### Features / Enhancements ###
 
+* Added in the `ColoradoHydroBaseRestDatastore` class to read in data from the
+DWR REST web services. A graph file can now use a TSID with the TSID location and
+`ColoradoHydroBaseRest` for the datastore.
 * Made Chart & Heatmap dialogs resizable again.
 
 # 3.6.1 #